fix: move database types to supabase folder
